requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
alabaster=0.7.13=pypi_0
albumentations=1.3.1=pypi_0
anyio=3.7.1=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
astor=0.8.1=pypi_0
astunparse=1.6.3=pypi_0
attrs=23.2.0=pypi_0
babel=2.14.0=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.12.3=pypi_0
blas=1.0=mkl
bleach=6.0.0=pypi_0
bottleneck=1.3.5=py37h7deecbd_0
ca-certificates=2024.3.11=h06a4308_0
cachetools=4.2.4=pypi_0
certifi=2022.12.7=py37h06a4308_0
cffi=1.15.1=pypi_0
charset-normalizer=3.3.2=pypi_0
comm=0.1.4=pypi_0
cycler=0.11.0=pypi_0
cython=3.0.10=pypi_0
debugpy=1.5.1=py37h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pypi_0
docutils=0.19=pypi_0
efficientnet=1.0.0=pypi_0
entrypoints=0.4=py37h06a4308_0
exceptiongroup=1.2.1=pypi_0
fastjsonschema=2.19.1=pypi_0
fftw=3.3.9=h5eee18b_2
fonttools=4.38.0=pypi_0
gast=0.3.3=pypi_0
google-auth=1.35.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.62.2=pypi_0
h5py=2.10.0=pypi_0
idna=3.7=pypi_0
image-classifiers=1.0.0=pypi_0
imageio=2.31.2=pypi_0
imagesize=1.4.1=pypi_0
imgaug=0.4.0=pypi_0
importlib-metadata=6.7.0=pypi_0
importlib-resources=5.12.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.15.2=py37h06a4308_0
ipyparallel=8.6.1=pypi_0
ipython=7.31.1=py37h06a4308_1
ipython-genutils=0.2.0=pypi_0
ipywidgets=8.1.2=pypi_0
jedi=0.18.1=py37h06a4308_1
jinja2=3.1.3=pypi_0
joblib=1.1.1=py37h06a4308_0
jsonschema=4.17.3=pypi_0
jupyter-core=4.12.0=pypi_0
jupyter-server=1.24.0=pypi_0
jupyter_client=7.4.9=py37h06a4308_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-widgets=3.0.10=pypi_0
keras=2.3.1=pypi_0
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.4.5=pypi_0
libedit=3.1.20230828=h5eee18b_0
libffi=3.2.1=hf484d3e_1007
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libprotobuf=3.20.3=he621ea3_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
markdown=3.4.4=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.5.3=pypi_0
matplotlib-inline=0.1.6=py37h06a4308_0
mistune=3.0.2=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
nbclassic=1.0.0=pypi_0
nbclient=0.7.4=pypi_0
nbconvert=7.6.0=pypi_0
nbformat=5.8.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=py37h06a4308_0
networkx=2.6.3=pypi_0
nose=1.3.7=pypi_0
notebook=6.5.6=pypi_0
notebook-shim=0.2.4=pypi_0
numexpr=2.8.4=py37he184ba9_0
numpy=1.21.6=pypi_0
numpy-base=1.21.5=py37ha15fc14_3
oauthlib=3.2.2=pypi_0
opencv-python=4.9.0.80=pypi_0
opencv-python-headless=4.9.0.80=pypi_0
openssl=1.1.1w=h7f8727e_0
opt-einsum=3.3.0=pypi_0
packaging=22.0=py37h06a4308_0
pandas=1.3.5=py37h8c16a72_0
pandocfilters=1.5.1=pypi_0
parso=0.8.3=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.5.0=pypi_0
pip=22.3.1=py37h06a4308_0
pkgutil-resolve-name=1.3.10=pypi_0
prometheus-client=0.17.1=pypi_0
prompt-toolkit=3.0.36=py37h06a4308_0
protobuf=4.24.4=pypi_0
psutil=5.9.0=py37h5eee18b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pypi_0
pygments=2.17.2=pypi_0
pyparsing=3.1.2=pypi_0
pyrsistent=0.19.3=pypi_0
python=3.7.3=h0371630_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytz=2024.1=pypi_0
pywavelets=1.3.0=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=23.2.0=py37h6a678d5_0
qtconsole=5.4.4=pypi_0
qtpy=2.4.1=pypi_0
qudida=0.0.4=pypi_0
readline=7.0=h7b6447c_5
requests=2.31.0=pypi_0
requests-oauthlib=2.0.0=pypi_0
rsa=4.9=pypi_0
scikit-image=0.16.2=pypi_0
scikit-learn=1.0.2=py37h51133e4_1
scipy=1.4.1=pypi_0
segmentation-models=1.0.1=pypi_0
send2trash=1.8.3=pypi_0
setuptools=65.6.3=py37h06a4308_0
shapely=2.0.4=pypi_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.3.1=pypi_0
snowballstemmer=2.2.0=pypi_0
soupsieve=2.4.1=pypi_0
sphinx=5.3.0=pypi_0
sphinxcontrib-applehelp=1.0.2=pypi_0
sphinxcontrib-devhelp=1.0.2=pypi_0
sphinxcontrib-htmlhelp=2.0.0=pypi_0
sphinxcontrib-jsmath=1.0.1=pypi_0
sphinxcontrib-qthelp=1.0.3=pypi_0
sphinxcontrib-serializinghtml=1.1.5=pypi_0
sqlite=3.33.0=h62c20be_0
tensorboard=2.2.2=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow-cpu=2.2.0=pypi_0
tensorflow-estimator=2.2.0=pypi_0
termcolor=2.3.0=pypi_0
terminado=0.17.1=pypi_0
testpath=0.6.0=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h1ccaba5_0
tornado=6.2=py37h5eee18b_0
tqdm=4.66.2=pypi_0
traitlets=5.7.1=py37h06a4308_0
typing-extensions=4.7.1=pypi_0
urllib3=2.0.7=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=pypi_0
websocket-client=1.6.1=pypi_0
werkzeug=2.2.3=pypi_0
wheel=0.38.4=py37h06a4308_0
widgetsnbextension=4.0.10=pypi_0
wrapt=1.16.0=pypi_0
xz=5.4.6=h5eee18b_0
zeromq=4.3.5=h6a678d5_0
zipp=3.15.0=pypi_0
zlib=1.2.13=h5eee18b_0
